requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
argbind==0.3.9
asttokens==2.4.1
astunparse==1.6.3
audioread==3.0.1
av==12.2.0
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.34.144
botocore==1.34.144
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.5
confection==0.1.5
contourpy==1.2.1
cryptography==42.0.8
ctranslate2==3.24.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.2
decorator==5.1.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
distro==1.9.0
docstring_parser==0.16
einops==0.8.0
emoji==2.8.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.0.1
faster-whisper==1.0.3
ffmpeg==1.4
ffmpy==0.3.2
filelock==3.15.4
fire==0.6.0
flatbuffers==24.3.25
flatten-dict==0.4.2
fonttools==4.53.1
fsspec==2024.6.1
funcy==2.0
future==1.0.0
gast==0.4.0
gcloud==0.18.3
google-api-core==2.19.1
google-auth==2.32.0
google-auth-oauthlib==0.4.6
google-cloud-speech==2.26.1
google-pasta==0.2.0
googleapis-common-protos==1.63.2
googlesearch-python==1.2.4
grpcio==1.64.1
grpcio-status==1.64.1
gTTS==2.5.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
julius==0.2.7
jupyter_client==8.6.2
jupyter_core==5.7.2
jwcrypto==1.5.6
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.14
openai-whisper==20231117
opt-einsum==3.3.0
packaging==24.1
parler_tts @ git+https://github.com/huggingface/parler-tts.git@8b8c576e2dbdc29172e30be7d68fac9357cd92c5
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
playsound==1.2.2
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.0rc1
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycaw==20240210
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pyparsing==3.1.2
pypiwin32==223
pyreadline3==3.4.1
Pyrebase4==4.8.0
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pyttsx3==2.90
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
randomname==0.2.1
RealTimeTTS==0.4.2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
stanza==1.6.1
stream2sentence==0.2.3
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.0
tbb==2021.13.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
torch-stoi==0.2.1
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
typer==0.12.3
typing_extensions==4.12.2
urllib3==1.26.19
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0